+++ commitlint.config.js
@@ -5,6 +5,7 @@ module.exports = {
   "rules": {
     "scope-case": [0, "never"],
     "body-min-length": [2, "always", 3],
-    "type-enum": [2, "always", ['feat', 'fix', 'docs', 'style', 'refactor', 'perf', 'test', 'chore', 'revert', 'dependencies', 'release']]
+    "type-enum": [2, "always", ['feat', 'fix', 'docs', 'style', 'refactor', 'perf', 'test', 'chore', 'revert', 'dependencies', 'release']],
+    "subject-case": [0, "never"]
   }
 }