+++ packages/mongodb-memory-server-core/src/util/__tests__/MongoBinaryDownload.test.ts
@@ -1,10 +1,14 @@
 import { promises as fspromises } from 'fs';
 import md5file from 'md5-file';
+import * as mkdirp from 'mkdirp';
+import { DryMongoBinary } from '../DryMongoBinary';
 import MongoBinaryDownload from '../MongoBinaryDownload';
+import MongoBinaryDownloadUrl from '../MongoBinaryDownloadUrl';
 import { envName, ResolveConfigVariables } from '../resolveConfig';
 import * as utils from '../utils';
 
 jest.mock('md5-file');
+jest.mock('mkdirp');
 
 describe('MongoBinaryDownload', () => {
   afterEach(() => {
@@ -141,5 +145,118 @@ describe('MongoBinaryDownload', () => {
     const du = new MongoBinaryDownload({ downloadDir: '/', checkMD5: false });
     const result = await du.makeMD5check('', '');
     expect(result).toEqual(undefined);
+  });
+
+  it('should return the correct path to binary name (getPath)', async () => {
+    const downloadDir = '/path/to/downloadDir';
+    jest.spyOn(DryMongoBinary, 'generateOptions').mockResolvedValue({
+      arch: 'x64',
+      version: '4.0.20',
+      downloadDir: downloadDir,
+      systemBinary: '',
+      os: {
+        os: 'linux',
+        dist: 'ubuntu',
+        release: '14',
+      },
+    });
+
+    const du = new MongoBinaryDownload({ downloadDir });
+
+    // @ts-expect-error because "getPath" is "protected"
+    const path = await du.getPath();
+    expect(path).toEqual(`${downloadDir}/mongod-x64-ubuntu-4.0.20`);
+  });
+
+  it('should print the download progress (printDownloadProgress)', () => {
+    const version = '4.0.20';
+    process.stdout.isTTY = false;
+    jest.spyOn(console, 'log').mockImplementation(() => void 0);
+    jest.spyOn(process.stdout, 'write').mockImplementation(() => true);
+
+    const du = new MongoBinaryDownload({
+      downloadDir: '/',
+      checkMD5: false,
+      version,
+      platform: 'linux',
+    });
+    expect(du.dlProgress.current).toBe(0);
+    du.dlProgress.length = 5242880;
+    du.dlProgress.totalMb = Math.round((du.dlProgress.length / 1048576) * 10) / 10;
+
+    du.printDownloadProgress({ length: 1048576 });
+    expect(console.log).toHaveBeenCalledWith(
+      `Downloading MongoDB "${version}": 20% (1mb / ${du.dlProgress.totalMb}mb)\r`
+    );
+    expect(console.log).toBeCalledTimes(1);
+    expect(process.stdout.write).not.toHaveBeenCalled();
+
+    du.printDownloadProgress({ length: 10 });
+    expect(console.log).toHaveBeenCalledTimes(1);
+    expect(process.stdout.write).not.toHaveBeenCalled();
+    expect(du.dlProgress.current).toBe(1048576 + 10);
+
+    process.stdout.isTTY = true;
+    du.printDownloadProgress({ length: 0 }, true);
+    expect(console.log).toHaveBeenCalledTimes(1);
+    expect(process.stdout.write).toHaveBeenCalledWith(
+      // its still "20" and "1" because of rounding
+      `Downloading MongoDB "${version}": 20% (1mb / ${du.dlProgress.totalMb}mb)\r`
+    );
+    expect(process.stdout.write).toHaveBeenCalledTimes(1);
+  });
+
+  it('should directly return the path if already exists', async () => {
+    const binaryPath = '/path/to/binary';
+    jest.spyOn(utils, 'pathExists').mockResolvedValue(true);
+
+    const du = new MongoBinaryDownload({ downloadDir: '/' });
+    jest
+      .spyOn(du, 'startDownload')
+      .mockImplementation(() => fail('Expected this function not to be called'));
+    // @ts-expect-error because "getPath" is "protected"
+    jest.spyOn(du, 'getPath').mockResolvedValue(binaryPath);
+
+    const returnValue = await du.getMongodPath();
+    expect(returnValue).toEqual(binaryPath);
+    expect(du.startDownload).not.toHaveBeenCalled();
+    expect(
+      // @ts-expect-error because "getPath" is "protected"
+      du.getPath
+    ).toHaveBeenCalledTimes(1);
+  });
+
+  it('should return the mongodb archive path (startDownload)', async () => {
+    const downloadUrl = 'https://fastdl.mongodb.org/linux/mongod-something-something.tgz';
+    const archivePath = '/path/to/archive.tgz';
+    jest.spyOn(mkdirp, 'default').mockResolvedValue(void 0);
+    jest.spyOn(fspromises, 'access').mockResolvedValue(void 0);
+    jest.spyOn(MongoBinaryDownloadUrl.prototype, 'getDownloadUrl').mockResolvedValue(downloadUrl);
+
+    const du = new MongoBinaryDownload({ downloadDir: '/' });
+    jest.spyOn(du, 'download').mockResolvedValue(archivePath);
+    jest.spyOn(du, 'makeMD5check');
+
+    const returnValue = await du.startDownload();
+    expect(returnValue).toEqual(archivePath);
+    expect(du.makeMD5check).toHaveBeenCalledWith(`${downloadUrl}.md5`, archivePath);
+    expect(du.download).toHaveBeenCalledWith(downloadUrl);
+  });
+
+  it('should return the mongodb archive path (startDownload)', async () => {
+    const customError = new Error('custom fs error');
+    jest.spyOn(mkdirp, 'default').mockResolvedValue(void 0);
+    jest.spyOn(fspromises, 'access').mockRejectedValue(customError);
+    jest.spyOn(console, 'error').mockImplementation(() => void 0);
+
+    const du = new MongoBinaryDownload({ downloadDir: '/' });
+
+    try {
+      await du.startDownload();
+      fail('Expected "startDownload" to fail');
+    } catch (err) {
+      expect(err.message).toEqual(customError.message);
+      expect(console.error).toHaveBeenCalledTimes(1);
+    }
   });
 });