+++ packages/mongodb-memory-server-core/src/util/errors.ts
@@ -1,7 +1,7 @@
 import { isNullOrUndefined } from './utils';
 
 export class StateError extends Error {
-  constructor(public wantedStates: [string], public gotState: string) {
+  constructor(public wantedStates: string[], public gotState: string) {
     super(
       `Incorrect State for operation: "${gotState}", allowed States: "[${wantedStates.join(',')}]"`
     );